pyproject.toml
[project]
name = "purchasing-concierge-local"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "google-adk[a2a]>=1.16.0",
  "google-genai>=1.3.0",
  "uvicorn>=0.30.6",
  "python-dotenv>=1.0.1"
]
